Type and general AST code clarity improvement
