--- semantic/ASTBase.hpp
@@ -1,20 +1,0 @@
-#ifndef __ASTBASE_HPP__
-#define __ASTBASE_HPP__
-
-#include <iostream>
-
-class AST
-{
-public:
-    virtual ~AST() {}
-    virtual void printOn(std::ostream &out) const = 0;
-    virtual void sem() {}
-};
-
-inline std::ostream &operator<<(std::ostream &out, const AST &t)
-{
-    t.printOn(out);
-    return out;
-};
-
-#endif
