+++ semantic/ast.hpp
@@ -5,6 +5,7 @@
 #include <string>
 #include <vector>
 #include <variant>
+#include "ASTabstracts.hpp"
 #include "types.hpp"
 
 extern std::map<std::string, int> globals;
@@ -12,161 +13,6 @@ extern std::map<std::string, int> globals;
 class FuncDef;
 class LocalDef;
 class Local;
-
-class Expr : public AST
-{
-public:
-    virtual int eval() const = 0;
-    // void type_check(Types t)
-    // {
-    //     sem();
-    //     if(type != t) {
-    //         std::cerr << "Type mismatch" << std::endl;
-    //         exit(1);
-    //     }
-    // }
-
-protected:
-    Type *type;
-};
-
-class Cond : public AST
-{
-public:
-    virtual int eval() const = 0;
-};
-
-class Stmt : public AST
-{
-public:
-    virtual void run() const = 0;
-};
-
-class LValue : public Expr
-{
-public:
-    virtual int eval() const = 0;
-    virtual std::string getName() const {};
-};
-
-class Const : public Expr
-{
-public:
-    Const(int n) : num(n) { type = new Integer(); }
-    virtual void printOn(std::ostream &out) const override
-    {
-        out << "Const(" << num << ")";
-    }
-    virtual int eval() const override
-    {
-        return num;
-    }
-    virtual void sem() override {}
-
-private:
-    int num;
-};
-
-class ConstChar : public Expr
-{
-public:
-    ConstChar(char v) : var(v) { type = new Character(); }
-    virtual void printOn(std::ostream &out) const override
-    {
-        out << "ConstChar(" << var << ")";
-    }
-    virtual int eval() const override
-    {
-        // return var;
-    }
-    virtual void sem() override {}
-
-private:
-    char var;
-};
-
-class ConstStr : public LValue
-{
-public:
-    ConstStr(char *s) : var(s) { type = new Str(); }
-    virtual void printOn(std::ostream &out) const override
-    {
-        out << "ConstStr(" << var << ")";
-    }
-    virtual const char *getStr()
-    {
-        return var;
-    }
-    virtual std::string getName() const override
-    {
-        return std::string(var);
-    }
-    virtual int eval() const override
-    {
-    }
-    virtual void sem() override {}
-
-private:
-    const char *var;
-};
-
-class ArrayDim : public AST
-{
-public:
-    ArrayDim() : dim() {}
-    ~ArrayDim()
-    {
-        for (Const *num : dim)
-        {
-            delete num;
-        }
-    }
-    virtual void printOn(std::ostream &out) const override
-    {
-        if (unknown)
-            out << "[ ]";
-        for (auto num = dim.rbegin(); num != dim.rend(); ++num)
-        {
-            out << "[" << **num << "]";
-        }
-    }
-    bool isEmpty()
-    {
-        return dim.empty();
-    }
-    void specifyUnknownFirstDim()
-    {
-        unknown = true;
-    }
-    void append(Const *num)
-    {
-        dim.push_back(num);
-    }
-    // TODO: for semantic action check each num that is of the correct type
-private:
-    std::vector<Const *> dim;
-    bool unknown = false;
-};
-
-class Array : public Type
-{
-
-public:
-    Array(Type *t, ArrayDim *d) : type(t), dim(d) {}
-    ~Array()
-    {
-        delete type;
-        delete dim;
-    }
-    virtual void printOn(std::ostream &out) const override
-    {
-        out << *type << " array" << *dim;
-    }
-
-private:
-    Type *type;
-    ArrayDim *dim;
-};
 
 class Id : public LValue
 {
@@ -380,38 +226,6 @@ public:
 
 private:
     std::vector<Stmt *> stmt_list;
-};
-
-class RetType : public Type
-{
-public:
-    RetType() { dtype = new Nothing(); }
-    RetType(Type *t) : dtype(t) {}
-    ~RetType() { delete dtype; }
-    virtual void printOn(std::ostream &out) const override
-    {
-        out << "RetType(" << *dtype << ")";
-    }
-
-private:
-    Type *dtype;
-};
-
-class FParType : public Type
-{
-public:
-    FParType(Type *t) : type(t) {}
-    ~FParType()
-    {
-        delete type;
-    }
-    virtual void printOn(std::ostream &out) const override
-    {
-        out << "FParType(" << *type << ")";
-    }
-
-private:
-    Type *type;
 };
 
 class FParam : public AST
